added filter method to cartTable to remove items from table whose qunatites were 0

+++ src/store/actions/cart.js
@@ -16,9 +16,10 @@ export const removeFromCartUnsafe = productId => (
 )
 
 export const removeFromCart = productId => (dispatch, getState) => {
-    console.log("got here");
+    if (getState().cart.quantityById[productId] > 0 ) {
+        dispatch(removeFromCartUnsafe(productId))
+    }
     // debugger;
-    dispatch(removeFromCartUnsafe(productId))
 }
 
 export const addToCart = productId => (dispatch, getState) => {
